mlops2/cat-dog-fast-api/requirements.txt
fastapi
uvicorn
numpy
pydantic
requests
joblib
python-multipart
tensorflow-cpu==2.15.0
Pillow
prometheus-client
